Update teslajson.py
I can host malicious content on eviltesla.com and bypass your validation checks:
```
>>> "https://eviltesla.com".endswith(('teslamotors.com','tesla.com'))
True
```
Hence, I suggest you need to do a better input validation. One simple way to do this is by prefixing the naked domain with a dot, as shown in this PR.

However, IMHO, the baseurl for the API should be hardcoded into the module, not imported from a third party site. I do understand your rational for doing this, but it's a very insecure design.